requirements.txt
openai>=1.30.0
sentence-transformers>=3.0.0
chromadb>=0.5.0
streamlit>=1.35.0
requests>=2.32.0
tiktoken>=0.7.0
python-dotenv>=1.0.1
tenacity>=8.3.0